Accept a raw database URI in config
It's often useful to be able to directly configure the full URI for
the database, such as when one wants to use ident authentication with
PostgreSQL.

With this commit, we allow a `URI` parameter to be set in the database
section of the configuration.  If this is set it overrides the
parameters `Host`, `Name`, `User`, `Passwd`, `Path`, and `SSLMode`.

If a path is specified in the URI, such as for sqlite, the directories
leading to the path will not be created as they would if `Path` had
been specified normally.